requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
curl-cffi==0.12.0
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.116.1
feedparser==6.0.11
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.7.0
git-filter-repo==2.47.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
gradio==5.38.0
gradio-client==1.11.0
greenlet==3.2.3
groovy==0.1.2
groq==0.30.0
grpcio==1.73.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keybert==0.9.0
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.70
langchain-groq==0.3.6
langchain-ollama==0.3.5
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.8
loguru==0.7.3
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
multitasking==0.0.12
mypy-extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
ollama==0.5.1
onnxruntime==1.22.1
openai==1.96.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
posthog==5.4.0
propcache==0.3.2
protobuf==6.31.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pygments==2.19.2
pypdf==5.8.0
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.4
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
starlette==0.47.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.15.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.37.0
typer==0.16.0
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
win32-setctime==1.2.0
yarl==1.20.1
yfinance==0.2.65
zipp==3.23.0
zstandard==0.23.0
